==> analysis-condition/03-condition-model.R <==
# condition models
library(tidyverse)
library(sdmTMB)
library(ggsidekick)
# remotes::install_github("seananderson/ggeffects", ref = "sdmTMB")
devtools::load_all(".")

# load overall species list
source("analysis-condition/00-species-list.R")
#
# # # or override with custom subset
# species_list <- list(
# # "North Pacific Spiny Dogfish",
# # "Pacific Ocean Perch",
# # "Pacific Cod",
# # "Walleye Pollock",
# # "Sablefish",
# # "Lingcod",
# "Bocaccio"
# # "Canary Rockfish",
# # "Quillback Rockfish",
# # "Redbanded Rockfish",
# # "Redstripe Rockfish", #
# # "Rougheye/Blackspotted Rockfish Complex", #
# # "Silvergray Rockfish", #
# # "Shortspine Thornyhead",
# # "Widow Rockfish", #
# # "Yelloweye Rockfish",
# # "Yellowmouth Rockfish", #
# # "Yellowtail Rockfish",
# # "Shortraker Rockfish",
# # "Rosethorn Rockfish",
# # "Harlequin Rockfish",
# # "Pygmy Rockfish",
# # "Sharpchin Rockfish",
# # "Darkblotched Rockfish",
# # "Greenstriped Rockfish",
# # "Petrale Sole", #
# # "Arrowtooth Flounder", #
# # "English Sole",#
# # "Dover Sole",#
# # "Rex Sole", #
# # "Flathead Sole",#
# # "Southern Rock Sole",#
# # "Slender Sole",#
# # "Pacific Sanddab",#
# # "Pacific Halibut",#
# # "Butter Sole",
# # "Pacific Hake",#
# # "Pacific Tomcod",
# # "Spotted Ratfish",
# # "Longnose Skate",
# # "Big Skate",
# # "Sandpaper Skate"
# )

index_list <- expand.grid(species = species_list,
                          maturity = c(
                                       # "imm",
                                       "mat"),
                          males = c(TRUE, FALSE)
                          ) %>%
  mutate(
    females = ifelse(males == FALSE & maturity == "mat", TRUE, FALSE),
    males = ifelse(maturity == "imm", FALSE, males)
  ) %>%
  distinct()


# Function for generating condition indices ----

calc_condition_indices <- function(species, maturity, males, females) {

  # species <- "Rex Sole"
  # maturity <- "mat"
  # males <- FALSE
  # females <- TRUE
  # stop_early <- TRUE
  stop_early <- FALSE

  # add_density <- FALSE
  add_density <- TRUE

  # # probably don't need these anymore
  unweighted <- TRUE
  # unweighted <- FALSE
  # adjusted_weights <- TRUE
  # adjusted_weights <- FALSE

  mat_threshold <- 0.5
  # knot_distance <- 5
  # knot_distance <- 10
  # knot_distance <- 15
  knot_distance <- 20
  fig_height <- 4 * 2
  fig_width <- 5 * 2
  delta_dens_model <- TRUE

  dens_model_total <- "dln-all-mar-2024" # this is for total
  dens_model_name1 <- "dln-all-shrunk" # these are `all catches' models
  dens_model_name2 <- "dln-only-sampled-shrunk" # these are `sampled catches' models

  theme_set(ggsidekick:::theme_sleek())

  spp <- gsub(" ", "-", gsub("\\/", "-", tolower(species)))

  mat_class <- maturity
  just_males <- males
  just_females <- females


  # browser()

  print(paste(species, maturity, "(males:", just_males, ")"))


  # Check which density models to use ----
  id0 <- readRDS(paste0("data-generated/density-index/",
                        dens_model_total, "/total/i-",
                        spp, "-total-", dens_model_total, "-",
                        knot_distance,
                        "-km.rds"))

  if1 <- readRDS(paste0("data-generated/density-index/",
                        dens_model_name1, "/mat-fem/i-",
                        spp, "-mat-fem-", dens_model_name1, "-",
                        knot_distance,
                        "-km.rds"))

  if2 <- readRDS(paste0("data-generated/density-index/",
                        dens_model_name2, "/mat-fem/i-",
                        spp, "-mat-fem-", dens_model_name2, "-",
                        knot_distance,
                        "-km.rds"))
  im1 <- readRDS(paste0("data-generated/density-index/",
                        dens_model_name1, "/mat-m/i-",
                        spp, "-mat-m-", dens_model_name1, "-",
                        knot_distance,
                        "-km.rds"))

  im2 <- readRDS(paste0("data-generated/density-index/",
                        dens_model_name2, "/mat-m/i-",
                        spp, "-mat-m-", dens_model_name2, "-",
                        knot_distance,
                        "-km.rds"))

  # in case imm isn't available
  ii1 <- NULL
  ii2 <- NULL
  try(ii1 <- readRDS(paste0("data-generated/density-index/",
                            dens_model_name1, "/imm/i-",
                            spp, "-imm-", dens_model_name1, "-",
                            knot_distance,
                            "-km.rds")))

  try(ii2 <- readRDS(paste0("data-generated/density-index/",
                            dens_model_name2, "/imm/i-",
                            spp, "-imm-", dens_model_name2, "-",
                            knot_distance,
                            "-km.rds")))

  # id1 <- bind_rows(if1,im1,ii1) %>% group_by(year, model_string) %>% summarise(est = sum(est))
  # id2 <- bind_rows(if2,im2,ii2) %>% group_by(year, model_string) %>% summarise(est = sum(est))
  #
  # ggplot(id0, aes(year, est)) + geom_line() +
  #   geom_line(data = id1, colour= "blue") +
  #   geom_line(data = id2, colour="red")

  isplit <- bind_rows(if1,im1,ii1) %>% bind_rows(., if2,im2,ii2) %>%
    group_by(year, model_string) %>% summarise(combined = sum(est))

  id0 <- id0 %>% select(-model_string)
  iall <- left_join(isplit, id0)

  iall <- iall %>% mutate(diff = combined - est,
                          lwr_diff = combined - lwr,
                          too_low = ifelse(lwr_diff < 0, lwr_diff, 0),
                          upr_diff = combined - upr,
                          too_high = ifelse(upr_diff > 0, upr_diff, 0)
  )

  ggplot(iall, aes(year, combined, colour = model_string)) + geom_line() +
    geom_line(aes(year, est), colour = "black")

  itest <- iall %>% group_by(species, model_string) %>%
    summarise(total_diff = sum(abs(diff)),
              mean_est = mean(est),
              sum_est = sum(est),
              mean_diff = mean(diff/est),
              prop_diff = total_diff/sum(est),
              too_low = sum(too_low),
              too_high = sum(too_high),
              total_ci_error = sum(abs(too_low)+too_high),
              prop_ci_error = total_ci_error/sum_est
    )

  itest$model_total <- dens_model_total

  dir.create(paste0("data-generated/compare-models/"),
             showWarnings = FALSE)
  saveRDS(itest, paste0("data-generated/compare-models/",
                        spp, "-relative-to-", dens_model_total, "-",
                        knot_distance,
                        "-km.rds"))

  itest <- itest %>%
    filter(prop_ci_error < 0.05)

  if(nrow(itest) == 1){
    dens_model_name <- itest$model_string
  } else {
    if(nrow(itest) == 0) { return(paste(species, "doesn't have split indices",
                                        "that cumulatively fall within 5% of total biomass CIs."))
    } else {
      dens_model_name <- itest[itest$total_diff==min(itest$total_diff),]$model_string
    }
  }

  # browser()

  if(stop_early) {
    return(itest)
  }

  # Load condition data and attach lagged density estimates ----
  dir.create(paste0("data-generated/condition-data-w-dens/"),
             showWarnings = FALSE)

  f <- paste0("data-generated/condition-data-w-dens/",
              spp, "-", mat_class, "-condition-best-model.rds")

  if (!file.exists(f)) {
    ds <- readRDS(paste0("data-generated/condition-data-black-swan/",
                         spp, "-mat-", mat_threshold, "-condition.rds")) %>%
      ungroup() %>%
      mutate(
        log_depth_c = log_depth - 5,
        DOY = as.numeric(strftime(time_deployed, format = "%j")),
        days_to_solstice = DOY - 172
      )

    # ds$year_smooth <- ds$year
    ds <- ds %>% filter(!is.na(depth_m))
    ds <- ds %>% filter(!is.na(days_to_solstice))
    ds <- ds %>% filter(!is.na(latitude))
    ds <- ds %>% filter(!is.na(longitude))

    ds$X <- NULL
    ds$Y <- NULL
    ds2 <- sdmTMB::add_utm_columns(ds,
                                   ll_names = c("longitude", "latitude"),
                                   utm_crs = 32609)

    nd0 <- ds2 %>%
      select(year, survey_abbrev, fishing_event_id, X, Y, log_depth) %>%
      mutate(
        survey_group = "TRAWL",
        survey_type = "SYN",
        days_to_solstice = 0,
        log_depth_c = log_depth - 5
      ) # can't predict prior to first year of density model

#
#     nd1 <- ds2 %>%
#       select(year, survey_abbrev, fishing_event_id, X, Y, log_depth) %>%
#       mutate(
#         survey_group = "TRAWL",
#         survey_type = "SYN",
#         days_to_solstice = 0,
#         log_depth_c = log_depth - 5,
#         year_true = year,
#         year_density = year - 1,
#         year = year_density
#       ) %>%
#       filter(year > 2000)
#     # can't predict prior to first year of density model


    if(mat_class == "mat") {

      mdf <- readRDS(paste0("data-generated/density-models/",
                            dens_model_name,"/mat-fem/",
                           spp, "-mat-fem-", dens_model_name, "-",
                           knot_distance,
                           "-km.rds"))
      mdm <- readRDS(paste0("data-generated/density-models/",
                            dens_model_name,"/mat-m/",
                           spp, "-mat-m-", dens_model_name, "-",
                           knot_distance,
                           "-km.rds"))

      mdf <- sdmTMB:::update_version(mdf)
      mdm <- sdmTMB:::update_version(mdm)

      nd0 <- nd0 %>% filter(year >= max(min(mdf$data$year),min(mdm$data$year)))
      # can't predict prior to first year of density model

      pd0f <- predict(mdf, newdata = nd0)
      pd0m <- predict(mdm, newdata = nd0)

      if (length(mdf$family$family)>1) {
        pd0f <- pd0f %>% mutate(
          density = mdf$family[[1]]$linkinv(est1) * mdf$family[[2]]$linkinv(est2)
        )
      } else {
        pd0f <- pd0f %>% mutate(density = exp(est))
      }

      if (length(mdm$family$family)>1) {
        pd0m <- pd0m %>% mutate(
          density = mdm$family[[1]]$linkinv(est1) * mdm$family[[2]]$linkinv(est2)
        )
      } else {
        pd0m <- pd0m %>% mutate(density = exp(est))
      }

      pd0 <- pd0f
      pd0$density <- pd0f$density + pd0m$density

    } else{

      dens_model_total <- dens_model_name

      fmi <- paste0("data-generated/density-models/",
                    dens_model_name,"/imm/",
                    spp, "-imm-", dens_model_name, "-",
                    knot_distance,
                    "-km.rds")

      if(file.exists(fmi)){
      md <- readRDS(fmi)
      } else {
        return(NA)
      }

      md <- sdmTMB:::update_version(md)


      nd0 <- nd0 %>% filter(year >= min(md$data$year))
      # can't predict prior to first year of density model

      pd0 <- predict(md, newdata = nd0)

      if (length(md$family$family)>1) {
        pd0 <- pd0 %>% mutate(
          density = md$family[[1]]$linkinv(est1) * md$family[[2]]$linkinv(est2)
        )
      } else {
        pd0 <- pd0 %>% mutate(density = exp(est))
      }

    }
    # browser()


    pd0 <- pd0 %>%
      mutate(
        log_density = log(density),
        model = dens_model_name
      ) %>%
      select(
        year, survey_abbrev, fishing_event_id, X, Y, log_depth,
        density, log_density, model
      ) %>%
      distinct()

    # pd1 <- pd1 %>%
    #   mutate(
    #     year = year_true,
    #     log_density_lag1 = log(density_lag1),
    #     model = dens_model_name2
    #   ) %>%
    #   select(
    #     year, survey_abbrev, fishing_event_id, X, Y, log_depth,
    #     year_density, density_lag1, log_density_lag1, model
    #   ) %>%
    #   distinct()

    d2 <- left_join(ds2, pd0) #%>% left_join(pd1)
    saveRDS(d2, f)
  } else {
    d2 <- readRDS(f)
  }



  # Select relevant data and grid ----

  # browser()

  if (mat_class == "mat") {
    if (just_males) {

      pf <- paste0(
        "data-generated/density-predictions/p-", spp,
        "-mat-m-", dens_model_name, "-", knot_distance,  "-km.rds"
      )
      if (file.exists(pf)) {
        d <- d2 %>% filter(group_name == "Mature males")
        group_tag <- "mat-m"
        group_label <- "mature males"

        # get current year density to scale condition index with
        gridA <- readRDS(pf) %>%
          select(year, X, Y, survey, depth, days_to_solstice, log_depth, density) %>%
          group_by(year) %>%
          mutate(
            sum_density = sum(density),
            prop_density = density / sum_density
          ) %>%
          ungroup() %>%
          group_by(year, survey) %>%
          mutate(
            survey_density = sum(density),
            prop_density_by_survey = density / survey_density
          )
      } else {
        print(paste("No density predictions for mature male ", species, "densities."))
        return(NA)
      }
    } else {
      if (just_females) {
        pf <- paste0(
          "data-generated/density-predictions/p-", spp,
          "-mat-fem-", dens_model_name, "-", knot_distance,  "-km.rds"
        )
        if (file.exists(pf)) {
          d <- d2 %>% filter(group_name == "Mature females")
          group_tag <- "mat-fem"
          group_label <- "mature females"

          # get current year density to scale condition index with
          gridA <- readRDS(pf) %>%
            select(year, X, Y, survey, depth, days_to_solstice, log_depth, density) %>%
            group_by(year) %>%
            mutate(
              sum_density = sum(density),
              prop_density = density / sum_density
            ) %>%
            ungroup() %>%
            group_by(year, survey) %>%
            mutate(
              survey_density = sum(density),
              prop_density_by_survey = density / survey_density
            )
        } else {
          print(paste("No density predictions for mature female ", species, "densities."))
          return(NA)
        }
      } else {
        pf <- paste0(
          "data-generated/density-predictions/p-", spp,
          "-all-mat-", dens_model_name, "-", knot_distance,  "-km.rds"
        )
        if (file.exists(pf)) {
          d <- d2 %>%
            filter(group_name %in% c("Mature females", "Mature males")) %>%
            rename(group_catch_weight_split = group_catch_weight)
          d3 <- d %>%
            group_by(fishing_event_id, group_name) %>%
            select(fishing_event_id, group_name, group_catch_weight_split) %>%
            distinct() %>%
            ungroup() %>%
            group_by(fishing_event_id) %>%
            summarise(group_catch_weight = sum(group_catch_weight_split))
          d <- left_join(d, d3)
          group_tag <- "mat"
          group_label <- "mature (females and males)"

          # get current year density to scale condition index with
          gridA <- readRDS(pf) %>%
            select(year, X, Y, survey, depth, days_to_solstice, log_depth, density) %>%
            group_by(year) %>%
            mutate(sum_density = sum(density),
                   prop_density = density / sum_density
            ) %>%
            ungroup() %>%
            group_by(year, survey) %>%
            mutate(survey_density = sum(density), prop_density_by_survey = density / survey_density)
        } else {
          print(paste("No density predictions for all mature ", species, "densities."))
          return(NA)
        }
      }
    }
  } else {
    if (mat_class == "imm") {
      pf <- paste0(
        "data-generated/density-predictions/p-", spp,
        "-imm-", dens_model_name, "-", knot_distance,  "-km.rds"
      )
      if (file.exists(pf)) {
        d <- d2 %>% filter(group_name %in% c("Immature"))
        group_tag <- "imm"
        group_label <- "immatures"

        # get current year density to scale condition index with
        gridA <- readRDS(pf) %>%
          select(year, X, Y, survey, depth, days_to_solstice, log_depth, density) %>%
          group_by(year) %>%
          mutate(
            sum_density = sum(density),
            prop_density = density / sum_density#,
            # log_density = log(density),
            # ann_log_density = mean(log_density)
          ) %>%
          ungroup() %>%
          # mutate(dens_dev = log_density - ann_log_density) %>%
          group_by(year, survey) %>%
          mutate(
            survey_density = sum(density),
            prop_density_by_survey = density / survey_density
          )

        # browser()
      } else {
        print(paste("No density predictions for immature ", species, "densities."))
        return(NA)
      }
    } else {
      pf <- paste0(
        "data-generated/density-predictions/p-", spp,
        "-total-", dens_model_total, "-", knot_distance,  "-km.rds"
      )
      if (file.exists(pf)) {
        # model everything together
        d <- d2

        # get current year density to scale condition index with
        gridA <- readRDS(pf) %>%
          select(year, X, Y, survey, depth, days_to_solstice, log_depth, density) %>%
          group_by(year) %>%
        mutate(sum_density = sum(density),
               prop_density = density / sum_density#,
               # log_density = log(density),
               # ann_log_density = mean(log_density)
        ) %>%
          ungroup() %>%
          # mutate(dens_dev = log_density - ann_log_density
          #      ) %>%
          group_by(year, survey) %>%
          mutate(survey_density = sum(density),
                 prop_density_by_survey = density / survey_density)
        ungroup()
      } else {
        print(paste("No density predictions for total ", species, "densities."))
        return(NA)
      }
    }
  }

  # Add density covariates to grid ----

  if (add_density) {

    if(mat_class == "mat") {
    # load lagged density predictions for full survey grid if going to be used as covariates condition
    gridBf <- readRDS(paste0(
      "data-generated/density-predictions/p-", spp,
      "-mat-fem-", dens_model_name, "-", knot_distance,  "-km.rds"
    )) %>%
      select(year, X, Y, survey, depth, days_to_solstice, log_depth, density) %>%
      rename(
        density_f = density
      )

    gridBm <- readRDS(paste0(
      "data-generated/density-predictions/p-", spp,
      "-mat-m-", dens_model_name, "-", knot_distance,  "-km.rds"
    )) %>%
      select(year, X, Y, survey, depth, days_to_solstice, log_depth, density) %>%
      rename(
        density_m = density
      )

    gridB <- full_join(gridBf, gridBm) %>% mutate(density = density_f + density_m)

    # browser()

    gridB <- gridB |>
      select(year, X, Y, survey, depth, days_to_solstice, log_depth, density, density_f, density_m) |>
      group_by(year) |>
      mutate(#sum_density = sum(density),
             #prop_density = density / sum_density,
             sum_density_f = sum(density_f),
             prop_density_f = density_f / sum_density_f,
             sum_density_m = sum(density_m),
             prop_density_m = density_m / sum_density_m
      ) |>
      ungroup() |>
      group_by(X, Y) |>
      mutate(
        log_density = log(density),
        cell_mean_density = mean(density),
        log_mean_density = log(cell_mean_density)
      ) |>
      ungroup()

    } else{

      gridB <- readRDS(paste0(
        "data-generated/density-predictions/p-", spp,
        "-imm-", dens_model_name, "-", knot_distance,  "-km.rds"
      )) |>
        select(year, X, Y, survey, depth, days_to_solstice, log_depth, density) |>
        group_by(year) |>
        mutate(sum_density = sum(density),
               prop_density = density / sum_density
        ) |>
        ungroup() |>
      group_by(X, Y) |>
        mutate(
          log_density = log(density),
          cell_mean_density = mean(density),
          log_mean_density = log(cell_mean_density)
        ) |>
        ungroup()
    }
  }


  # Make mesh ----
  hist(d$cond_fac)
  hist(log(d$cond_fac))

  d <- d %>% mutate(
    survey_group = as.factor(
      case_when(
        survey_abbrev %in% c("HBLL OUT N", "HBLL OUT S")~"HBLL",
        survey_abbrev %in% c("MSSM QCS", "MSSM WCVI")~"MSSM",
        survey_abbrev %in% c("OTHER", "HS MSA", "SYN HS", "SYN QCS", "SYN WCHG", "SYN WCVI")~"TRAWL",
        survey_series_id == 68~"HAKE",
        TRUE~survey_abbrev
      ))
    )
  d <- d %>% filter(year >= min(gridA$year))

  mesh <- make_mesh(d, c("X", "Y"), cutoff = knot_distance)

  ggplot() +
    inlabru::gg(mesh$mesh) +
    coord_fixed() +
    geom_point(aes(X, Y, size = total_weight), data = d2) +
    geom_point(aes(X, Y, colour = group_catch_weight), data = d) +
    facet_wrap(~year) +
    scale_color_viridis_c(trans = ggsidekick::fourth_root_power_trans())

  ggplot() +
    inlabru::gg(mesh$mesh) +
    coord_fixed() +
    geom_point(aes(X, Y, size = group_catch_weight), data = d2) +
    geom_point(aes(X, Y, colour = log(cond_fac)), size = 0.5, data = d) +
    facet_wrap(~year) +
    scale_color_gradient2()

  # load refine function
  refine_cond_model <- function(m, set_formula = cond_formula, dist = knot_distance) {
    s <- sanity(m)
    t <- tidy(m, "ran_pars", conf.int = TRUE)
    if (!all(s) & !m$call$share_range) {
      # browser()
      t <- tidy(m, "ran_pars", conf.int = TRUE)
      if (abs(diff(t$estimate[t$term == "range"])) < dist|!s$range_ok) {
        m <- update(m,
          formula = set_formula,
          # formula = as.list(m[["formula"]]),
          share_range = TRUE,
          weights = m$data$sample_multiplier,
          # # spatial = as.list(m[["spatial"]]),
          # # spatiotemporal = as.list(m[["spatiotemporal"]]),
          # # extra_time = m$extra_time,
          # # family = m$family,
          data = m$data, mesh = m$spde
        )
        s <- sanity(m)
        t <- tidy(m, "ran_pars", conf.int = TRUE)
      }
    }
    if (!all(s)) {
      if (t$estimate[t$term == "sigma_O"] < 0.001) {
        m <- update(m,
          formula = set_formula,
          weights = m$data$sample_multiplier,
          spatial = "off",
          data = m$data, mesh = m$spde
        )
      }
    }
    sanity(m)
    return(m)
  }

  # Estimate condition model ----
  d %>%
    group_by(survey_group) %>%
    summarise(n = n())

    d$sample_multiplier <- 1
    model_name <- "apr-2024"

  dir.create(paste0("data-generated/condition-models-",
                    group_tag, "/", model_name, "/"),
             showWarnings = FALSE)
  dir.create(paste0("data-generated/cond-index/", model_name, "/"),
             showWarnings = FALSE)


  mf <- paste0(
    "data-generated/condition-models-", group_tag, "/", model_name, "/", spp, "-c-",
    group_tag, "-", model_name, "-", knot_distance, "-km.rds"
  )

  if (diff(range(d$days_to_solstice))<60) {
    if (length(unique(d$survey_group)) == 1) {
    cond_formula <- cond_fac ~ 1 + days_to_solstice
    } else {
    cond_formula <- cond_fac ~ survey_group + days_to_solstice
    }
  } else {
    if (length(unique(d$survey_group)) == 1) {
    cond_formula <- cond_fac ~ 1 + poly(days_to_solstice, 2)
    } else {
    cond_formula <- cond_fac ~ survey_group +
      poly(days_to_solstice, 2)
    }
  }

  # if commented out, will be forced to rerun model
  if (file.exists(mf)) {
    try(m <- readRDS(mf))
  }

  # browser()

  if (!exists("m")) {

    # TODO: Add DOY with cyclic smoother if we add in commercial data?
    # require(mgcv)
    # set.seed(6)
    # x <- sort(runif(200)*10)
    # z <- runif(200)
    # f <- sin(x*2*pi/10)+.5
    # y <- rpois(exp(f),exp(f))
    #
    # ## finished simulating data, now fit model...
    # b <- gam(y ~ s(x,bs="cc",k=12) + s(z),family=poisson,
    #          knots=list(x=seq(0,10,length=12)))
    # ## or more simply
    # b <- gam(y ~ s(x,bs="cc",k=12) + s(z),family=poisson,
    #          knots=list(x=c(0,10)))
    #
    # ## plot results...
    # par(mfrow=c(2,2))
    # plot(x,y);plot(b,select=1,shade=TRUE);lines(x,f-mean(f),col=2)
    # plot(b,select=2,shade=TRUE);plot(fitted(b),residuals(b))

    sort(unique(d$year))

    m <- sdmTMB(cond_formula,
      weights = d$sample_multiplier,
      mesh = mesh,
      data = d,
      spatial = "on",
      spatiotemporal = "rw",
      # spatiotemporal = "off",
      extra_time = sdmTMB:::find_missing_time(d$year),
      share_range = FALSE,
      silent = FALSE,
      time = "year",
      # reml = TRUE,
      family = lognormal(link = "log"),
      # family = student(df = 5),
      # control = sdmTMBcontrol(newton_loops = 1L),
      priors = sdmTMBpriors(
        matern_s = pc_matern(range_gt = knot_distance,
                             sigma_lt = 2),
        matern_st = pc_matern(range_gt = knot_distance,
                              sigma_lt = 2)
      )
    )

    print(paste(species, maturity, "(males:", just_males, ")"))

    m <- refine_cond_model(m1, set_formula = cond_formula,
                            dist = knot_distance)
    saveRDS(m, mf)
  } else {

    ## make sure if converged last time
    # m1 <- refine_cond_model(m1, set_formula = cond_formula, dist = knot_distance)
    # saveRDS(m1, mf)
  }

  m
  m$sd_report
  tidy(m, "ran_pars", conf.int = TRUE)

  # m <- m1

  # Add density dependence ----
  ## don't do this for now, but can be used to explore utility of covariates
  if (add_density) {

    d$log_density_c <- d$log_density - mean(d$log_density, na.rm = TRUE)

    # if(unweighted) {
      d$sample_multiplier <- 1

      # model_name <- "apr-2024-density"
      # model_name <- "apr-2024-not-total-density"
      model_name <- "apr-2024-cell-means"

      # model_name <- "doy-ld0c" # next update, change to this
    # } else {
    #   if(adjusted_weights){
    #     model_name <- "doy-d0c-within-yr-weights"
    #   }
    # }

    dir.create(paste0("data-generated/condition-models-", group_tag,
                      "/", model_name, "/"), showWarnings = FALSE)

    dir.create(paste0("data-generated/cond-index/", model_name, "/"),
               showWarnings = FALSE)

    if (diff(range(d$days_to_solstice))<60) {
      if (length(unique(d$survey_group)) == 1) {
        cond_formula2 <- cond_fac ~ 1 +
          days_to_solstice + log_density_c
      } else {
        cond_formula2 <- cond_fac ~ survey_group +
          days_to_solstice + log_density_c
      }
    } else {
      if (length(unique(d$survey_group)) == 1) {
      cond_formula2 <- cond_fac ~ 1 +
        poly(days_to_solstice, 2) +
        log_density_c
        # poly(log_density_c, 2)
        # ann_log_density_c + dens_dev
        # poly(ann_log_density_c, 2) + poly(dens_dev, 2)
        # poly(log_density_lag1_c, 2)
    } else {
      cond_formula2 <- cond_fac ~ survey_group +
        poly(days_to_solstice, 2) +
        log_density_c
      # poly(log_density_c, 2)
        # ann_log_density_c + dens_dev
        # poly(ann_log_density_c, 2) + poly(dens_dev, 2)
        # # poly(log_density_lag1_c, 2)
    }
    }

    mesh2 <- make_mesh(d, c("X", "Y"), cutoff = 15)

    mf2 <- paste0(
      "data-generated/condition-models-", group_tag, "/", model_name, "/",
      spp, "-c-", group_tag, "-", model_name, "-", knot_distance, "-km.rds"
    )

    if (file.exists(mf2)) {
      try(m2 <- readRDS(mf2))
    }

    if (!exists("m2")) {

      m2 <- update(m, cond_formula2,
        weights = d$sample_multiplier,
        spatial = "on",
        spatiotemporal = "rw",
        share_range = FALSE,
        # control = sdmTMBcontrol(upper = list(b_j = bx)),
        mesh = mesh2,
        data = d
      )

      rm(m)

      saveRDS(m2, mf2)
      m2 <- refine_cond_model(m2, set_formula = cond_formula2,
                              # upper_limits = bx,
                              dist = knot_distance)
      saveRDS(m2, mf2)
    } else {
      # saveRDS(m2, mf2)
    }

    sanity(m2)
    m2
    m2$sd_report
    tidy(m2, "ran_pars", conf.int = TRUE)

    ## check if effect of density is negative by more than the SE on the estimate
    ## if not, revert to m1 for index generation
    t <- tidy(m2, conf.int = TRUE)

    if(t$estimate[t$term == "log_density_c"] < - t$std.error[t$term == "log_density_c"]){
      m <- m2
    } else {
      m <- readRDS(mf)
    }
  }


  # Filter grid ----
  if (add_density) {
    grid <- gridB |> mutate(
      log_density_c = log_mean_density - mean(d$log_density, na.rm = TRUE)
      )
    if(mat_class == "mat") {
      if (just_males) {
      grid <- grid |> mutate(prop_density = prop_density_m)
      }else {
      grid <- grid |> mutate(prop_density = prop_density_f)
      }
    }
  } else {
    grid <- gridA
  }

  grid$survey_group <- "TRAWL"

  sort(unique(m$data$year))
  sort(unique(grid$year))

  # might be redundant
  grid <- filter(grid, year %in% sort(union(unique(m$data$year), m$extra_time)))


  # Check R2

  # browser()
  # sanity(m)
  # m
  # m$sd_report
  # tidy(m, conf.int = TRUE)
  # tidy(m, "ran_pars", conf.int = TRUE)

  # (r2 <- r2.sdmTMB(m))
  #r2$r2[1]

  # Has this model been run before? ----

  i1 <- paste0(
    "data-generated/cond-index/", model_name,
    "/cond-index-", group_tag, "-", spp, "-",
    model_name, "-", knot_distance, "-km.rds"
  )

  # browser()
  if (!file.exists(i1)) {

  pc <- predict(m, newdata = grid, return_tmb_object = TRUE)


  p2 <- pc$data %>%
    # filter(!(year == 2020)) %>%
    mutate(cond = exp(est))


  # filter to plot only cells representing 99% of mean predicted biomass
  # cells must be defined by "X", "Y", time by "year", and biomass/abundance stored as "density"
  p2 <- trim_predictions_by_year(p2, 0.001)


  # Map model predictions ----

  dset <- readRDS(paste0("data-generated/density-data/", spp, ".rds")) %>%
    filter(year >= min(m$data$year), year <= max(m$data$year))

  unique(dset$survey_abbrev)

  # all sets
  set_list <- dset %>%
    select(fishing_event_id, longitude, latitude, X, Y, year, catch_weight, catch_count) %>%
    distinct() %>%
    mutate(
      fishing_event_id = as.factor(fishing_event_id),
      lon = longitude, lat = latitude)

  # # just sampled sets
  # set_list <- d2 %>% select(fishing_event_id, longitude, latitude, X, Y) %>% distinct() %>%
  #   mutate(lon = longitude, lat = latitude)

  model_dat <- d %>%
    group_by(fishing_event_id) %>%
    mutate(
      fishing_event_id = as.factor(fishing_event_id),
      count = n()
    )

  model_dat <- left_join(set_list, model_dat, multiple = "all") %>% mutate(
    density = group_catch_weight,
    caught = ifelse(catch_count > 0 | catch_weight > 0, 1, 0),
    count = ifelse(is.na(count), 0, count),
    present = ifelse(count > 0, 1, ifelse(caught == 1, 0, NA_integer_))
  )

  # model_dat %>% group_by(present, caught) %>% summarise(n = n()) %>% View()
  p2$log_cond <- log(p2$cond)
  p2 <- p2 %>% mutate(cond_trim = ifelse(cond > quantile(p2$cond, 0.99),
    quantile(p2$cond, 0.99), cond
  ))
  g <- gfenvirocor::plot_predictions(p2, model_dat, # extrapolate_depth = FALSE,
    # fill_column = "log_cond",
    fill_column = "cond_trim",
    fill_label = "Condition \nfactor",
    pt_column = "count",
    pt_label = "Fish \nsampled",
    pt_size_range = c(0.5, 4),
    pos_pt_fill = NA,
    bin_pt_col = "black",
    pos_pt_col = "red",
    # x_buffer = c(-0, 0),
    # y_buffer = c(-0, 0),
    fill_scale =
      ggplot2::scale_fill_viridis_c(),
    # ggplot2::scale_fill_viridis_c(trans = "log10"),
    bounds = grid,
    rotation_angle = 30, show_raw_data = TRUE
  )

  # browser()

  g <- g + facet_wrap(~year, ncol = 7) +
    ggtitle(paste0(species, ": ", group_label, " ", model_name))

  dir.create(paste0("figs/cond-", model_name, "/"), showWarnings = FALSE)
  ggsave(paste0("figs/cond-", model_name, "/condition-map-wide-", spp, "-", group_tag, "-",
                model_name, "-", knot_distance, "-km.png"),
         height = fig_height*1.5, width = fig_width
  )

}

  # Get coastwide index ----

  if (!file.exists(i1)) {

    ind2 <- get_index(pc, area = grid$prop_density, bias_correct = TRUE)
    ind2$species <- species
    ind2$group <- group_label
    ind2$model_string <- model_name
    ind2$dens_model_name <- dens_model_name

    saveRDS(ind2, i1)
  } else {
    ind2 <- readRDS(i1)
  #
  #   ind2$species <- species
  #   ind2$group <- group_label
  #   ind2$model_string <- model_name
  #
  #   saveRDS(ind2, i1)
  }

  ggplot(ind2, aes(year, est)) +
    geom_line() +
    geom_ribbon(aes(ymin = lwr, ymax = upr), alpha = 0.4) +
    xlab("Year") +
    ylab("Predicted average condition factor") +
    labs(title = paste0(species, ": ", group_label, " ", model_name)#,
         # subtitle = paste("conditional R2:", r2$r2[1], "marginal R2:", r2$r2[2])
         )

  ggsave(paste0("figs/cond-", model_name, "/condition-index-", spp, "-",
                group_tag, "-", model_name, "-", knot_distance, "-km.png"),
    height = fig_height / 2, width = fig_width / 2
  )


  # # Get survey-specific condition indices ----
  # i2 <- paste0("data-generated/cond-index/", model_name, "/survey-cond-indices-", spp, "-", group_tag, "-", model_name, "-", knot_distance, "-km.rds")
  #
  # if(!file.exists(i2)) {
  #
  # ## TODO: if we want to split by stock area, will need to edit grid first to add species specific stock areas
  # preds <- grid %>%
  #   split(.$survey) %>%
  #   lapply(function(x) predict(m, newdata = x, return_tmb_object = TRUE))
  #
  # inds <- purrr::map_dfr(preds, function(.x)
  #   get_index(.x, area = .x$data$prop_density_by_survey, bias_correct = TRUE), .id = "region")
  #
  # survey_years <- m$data %>% select(survey_abbrev, year) %>% distinct() %>%
  #   mutate(region = ifelse(survey_abbrev == "HS MSA", "SYN HS",
  #                   ifelse(survey_abbrev == "MSSM QCS", "SYN QCS",
  #                   ifelse(survey_abbrev == "MSSM WCVI", "SYN WCVI",
  #                          survey_abbrev))))
  #
  # ind3 <- left_join(survey_years, inds, multiple = "all") %>%
  #   filter(!is.na(est))
  #
  # saveRDS(ind3, i2)
  #
  # } else {
  # ind3 <- readRDS(i2)
  # }
  #
  # ggplot(ind3, aes(year, est, fill = region)) +
  #   geom_line(aes(colour = region)) +
  #   geom_ribbon(aes(ymin = lwr, ymax = upr), alpha = 0.1) +
  #   xlab("Year") +
  #   ylab("Predicted average condition factor") +
  #   labs(title = paste0(species, ": ", group_label, " ", "-", model_name))
  #
  # ggsave(paste0("figs/condition-index-", spp, "-split-", group_tag, "-", model_name, "-", knot_distance, "-km.png"),
  #        height = fig_height / 2, width = fig_width/1.5
  # )


  #   map(get_index(., area = .$data$prop_density, bias_correct = FALSE)) %>%
  #   bind_rows(.id = "survey")


  # saveRDS(p2, paste0("data-generated/cond-predictions/", spp,
  #  "-pc", "-", group_tag, "-", model_name, "-", knot_distance, "-km.rds"))
  # saveRDS(ind2, paste0("data-generated/cond-index/", model_name, "/", spp, "-pc",
  #  "-", group_tag, "-", model_name, "-", knot_distance, "-km.rds"))


  ## probably don't need these for every model but keeping here for interest
  #
  # ggplot(p2, aes(X, Y, colour = log_density, fill = log_density)) +
  #   geom_tile(width = 2, height = 2, alpha = 1) +
  #   # facet_wrap(~year) +
  #   scale_fill_viridis_c() +
  #   scale_colour_viridis_c()
  #
  ## ggsave(paste0("figs/density-map-", spp, "-", group_tag, "-", model_name,
  #   "-", knot_distance, "-km.png"), height = fig_height, width = fig_width)
  # ggplot(p2, aes(X, Y,
  #                colour = omega_s,
  #                fill = omega_s)) +
  #   geom_tile(width = 2, height = 2, alpha = 1) +
  #   # facet_wrap(~year) +
  #   scale_fill_gradient2() +
  #   scale_colour_gradient2()
  #
  ## ggsave(paste0("figs/condition-omega-map-", spp, "-", group_tag,
  ##  model_name, "-", knot_distance, "-km.png"), height = fig_height, width = fig_width)
  #
  #
  # ggplot(p2, aes(X, Y, colour = epsilon_st, fill = epsilon_st)) +
  #   geom_tile(width = 2, height = 2, alpha = 1) +
  #   facet_wrap(~year) +
  #   scale_fill_gradient2() +
  #   scale_colour_gradient2()
  #
  ## ggsave(paste0("figs/condition-epsilon-map-", spp, "-", group_tag,
  ##  model_name, "-", knot_distance, "-km.png"), height = fig_height, width = fig_width)


  # plot_covariates <- TRUE
  plot_covariates <- FALSE

  if (plot_covariates) {

    nd <- data.frame(
      days_to_solstice = seq(min(d$days_to_solstice),
                             max(d$days_to_solstice),
                             length.out = 50
      ),
      survey_group = "TRAWL",
      log_density_c = 0,
      ann_log_density_c = 0,
      dens_dev = 0,
      year = 2021L # a chosen year
    )
    pd <- predict(m, newdata = nd, se_fit = TRUE, re_form = NA)

    ggplot(pd, aes(days_to_solstice, exp(est),
                   ymin = exp(est - 1.96 * est_se),
                   ymax = exp(est + 1.96 * est_se)
    )) +
      geom_line() +
      geom_ribbon(alpha = 0.4) +
      scale_x_continuous() +
      coord_cartesian(expand = F) +
      labs(x = "Days to solstice", y = "condition") +
      ggtitle(paste0(species, ": ", group_label, " ", model_name))

    ggsave(paste0("figs/cond-", model_name, "/effect-plot-doy-", spp, "-",
                  group_tag, "-", model_name, "-", knot_distance, "-km.png"),
           height = fig_height / 2, width = fig_width / 2
    )

    # Effect plots ----
    if (add_density) {
      nd2 <- data.frame(
        days_to_solstice = 0,
        survey_group = "TRAWL",
        log_density_c = seq(min(d$log_density_c),
                            max(d$log_density_c),
                            length.out = 50
        ),
        ann_log_density_c = 0,
        # ann_log_density_c = seq(min(d$ann_log_density_c),
        #                         max(d$ann_log_density_c),
        #                         length.out = 50
        # ),
        dens_dev = 0,
        year = 2021L # a chosen year
      )
      pd <- predict(m, newdata = nd2, se_fit = TRUE, re_form = NA)

      if(model_name == "all-st2002-doy-lddev") {
        nd3 <- data.frame(
          days_to_solstice = 0,
          survey_group = "TRAWL",
          log_density_c = seq(min(d$log_density_c),
                              max(d$log_density_c),
                              length.out = 50
          ),
          ann_log_density_c = 0,
          # ann_log_density_c = seq(min(d$ann_log_density_c),
          #                         max(d$ann_log_density_c),
          #                         length.out = 50
          # ),
          dens_dev = 0,
          year = 2021L # a chosen year
        )
        pd <- predict(m, newdata = nd3, se_fit = TRUE, re_form = NA)

        ggplot(pd, aes(ann_log_density_c, exp(est),
                       ymin = exp(est - 1.96 * est_se),
                       ymax = exp(est + 1.96 * est_se)
        )) +
        geom_line() +
        geom_ribbon(alpha = 0.4) +
        scale_x_continuous() +
        coord_cartesian(expand = F) +
        labs(x = "ann_log_density_c", y = "condition") +
        ggtitle(paste0(species, ": ", group_label, " ", model_name))
    } else {
      ggplot(pd, aes(log_density_c, exp(est),
                     ymin = exp(est - 1.96 * est_se),
                     ymax = exp(est + 1.96 * est_se)
      )) +
        geom_line() +
        geom_ribbon(alpha = 0.4) +
        scale_x_continuous() +
        coord_cartesian(expand = F) +
        labs(x = "log_density_c", y = "condition") +
        ggtitle(paste0(species, ": ", group_label, " ", model_name))

    }

    ggsave(paste0("figs/cond-", model_name, "/effect-plot-density-", spp, "-",
                  group_tag, "-", model_name, "-", knot_distance, "-km.png"),
           height = fig_height / 2, width = fig_width / 2
    )

    }


    #
    # nd <- data.frame(
    #   days_to_solstice = seq(min(d$days_to_solstice),
    #                          max(d$days_to_solstice),
    #                          length.out = 50
    #   ),
    #   survey_group = "TRAWL",
    #   log_density_c = 0,
    #   ann_log_density_c = 0,
    #   dens_dev = 0,
    #   year = 2021L # a chosen year
    # )
    # pd <- predict(m, newdata = nd, se_fit = TRUE, re_form = NA)
    #
    # ggplot(pd, aes(days_to_solstice, exp(est),
    #                ymin = exp(est - 1.96 * est_se),
    #                ymax = exp(est + 1.96 * est_se)
    # )) +
    #   geom_line() +
    #   geom_ribbon(alpha = 0.4) +
    #   scale_x_continuous() +
    #   coord_cartesian(expand = F) +
    #   labs(x = "days_to_solstice", y = "condition")
    #

    # (v <- visreg::visreg(m1, xvar = "survey_group"))
    # (v <- visreg::visreg(m2, xvar = "log_density_c"))

    # g <- ggeffects::ggeffect(m, paste0("log_density_c [",
    #   range(d$log_density_c)[1], ":", range(d$log_density_c)[2], "by=0.05]"))
    # plot(g)


    # TODO: not been tested with recent model configurations

    # ggplot(p3, aes(X, Y, colour = (est_non_rf), fill = (est_non_rf))) +
    #   geom_tile(width = 2, height = 2, alpha = 1) +
    #   facet_wrap(~year) +
    #   scale_fill_gradient2() +
    #   scale_colour_gradient2()
    #
    # ggsave(paste0("figs/condition-dd-map-", spp, "-", model_name, "-",
    #               knot_distance, "-km.png"))
    #
    # g <- g2 <- g3 <- g4 <- pd <- NULL
    #
    # g <- ggeffects::ggeffect(m, paste0(
    #   "log_depth [",
    #   range(d$log_depth)[1], ":", range(d$log_depth)[2], "by=0.05]"
    # ))
    # plot(g)
    #
    # g2 <- ggeffects::ggeffect(m, paste0(
    #   "log_mean_density_lag1 [", range(d$log_mean_density_lag1)[1],
    #   ":", range(d$log_mean_density_lag1)[2], "by=0.1]"
    # ))
    # plot(g2)

    # g3 <- ggeffects::ggeffect(m, paste0(
    #   "dens_dev [", range(d$dens_dev)[1], ":",
    #   range(d$dens_dev)[2], "by=0.1]"
    # ))
    # plot(g3)
    #
    # g4 <- ggeffects::ggeffect(m, paste0(
    #   "log_density_lag1 [", range(d$log_density_lag1)[1], ":",
    #   range(d$log_density_lag1)[2], "by=0.1]"
    # ))
    # plot(g4)
    #
    #
    # ## if time-varying depth included
    # nd <- expand.grid(
    #   log_depth = seq(min(d$log_depth),
    #     max(d$log_depth),
    #     length.out = 100
    #   ),
    #   year = unique(d$year) # all years
    # )
    #
    # nd$log_density <- 0
    # nd$log_density_lag1 <- 0
    #
    # pd <- predict(m2, newdata = nd, se_fit = TRUE, re_form = NA)
    # pd <- pd %>%
    #   group_by(year) %>%
    #   mutate(max_est = max(est, na.rm = T),
    #          xintercept = log_depth[est == max_est])
    #
    #
    # ggplot(pd, aes(exp(log_depth), exp(est),
    #   ymin = exp(est - 1.96 * est_se),
    #   ymax = exp(est + 1.96 * est_se),
    #   group = as.factor(year)
    # )) +
    #   geom_vline(aes(
    #     xintercept = exp(xintercept),
    #     group = year, colour = year
    #   ), alpha = 0.4) +
    #   geom_line(aes(colour = year), lwd = 0.5) +
    #   geom_ribbon(aes(fill = year), alpha = 0.1) +
    #   scale_colour_viridis_c() +
    #   scale_fill_viridis_c() +
    #   # scale_y_continuous(trans = "sqrt") +
    #   coord_cartesian(
    #     expand = F,
    #     xlim = c(20, 300),
    #     ylim = c(NA, exp(max(pd$max_est)) + 0.1)
    #   ) +
    #   # coord_cartesian(expand = F, ylim = c(NA, 1)) +
    #   # facet_wrap(~year_bin) +
    #   geom_rug(data = filter(m2$data, catch_weight > 0),
    #            aes(depth_m), inherit.aes = FALSE, alpha = 0.02) +
    #   labs(x = "Depth (m)", y = "Condition factor") +
    #   # labs(x = "Depth (m)", y = "Biomass density (kg/km2)") +
    #   # ylab(paste(group, "biomass density")) +
    #   gfplot::theme_pbs()
    #

    # gg <- list(g, g2, g3, g4, pd)
    # dir.create(paste0("data-generated/cond-effects/"), showWarnings = FALSE)
    # saveRDS(gg, paste0("data-generated/cond-effects/", spp, "-", group_tag, "-", model_name, "-", knot_distance, "-km.rds"))


    ## SVC plots
    #
    # coefs <- tidy(m2, conf.int = TRUE)
    # log_density_coef <- coefs$estimate[coefs$term == "log_density"]
    #
    # ggplot(p3, aes(X, Y,
    #                colour = log_density_coef + zeta_s_log_density,
    #                fill = log_density_coef + zeta_s_log_density)) +
    #   geom_tile(width = 2, height = 2, alpha = 1) +
    #   # facet_wrap(~year) +
    #   scale_fill_gradient2() +
    #   scale_colour_gradient2()
    #
    #
    # dens_dev_coef <- coefs$estimate[coefs$term == "dens_dev"]
    #
    # ggplot(p3, aes(X, Y,
    #                colour = dens_dev_coef + zeta_s_dens_dev,
    #                fill = dens_dev_coef + zeta_s_dens_dev)) +
    #   geom_tile(width = 2, height = 2, alpha = 1) +
    #   # facet_wrap(~year) +
    #   scale_fill_gradient2() +
    #   scale_colour_gradient2()
    #
    # log_mean_density_lag1_coef <- coefs$estimate[coefs$term == "log_mean_density_lag1"]
    #
    # ggplot(p3, aes(X, Y,
    #                colour = log_mean_density_lag1_coef + zeta_s_log_mean_density_lag1,
    #                fill = log_mean_density_lag1_coef + zeta_s_log_mean_density_lag1)) +
    #   geom_tile(width = 2, height = 2, alpha = 1) +
    #   # facet_wrap(~year) +
    #   scale_fill_gradient2() +
    #   scale_colour_gradient2()
    #
    # ggsave(paste0("figs/condition-zeta-map-", spp, "-", group_tag, "-", model_name, "-", knot_distance, "-km.png"), height = fig_height, width = fig_width)
  }
}

# Run with pmap -----

# index_list <- index_list[2, ]
pmap(index_list, calc_condition_indices)

# Run with furrr ----

library(gfenvirocor)
library(future)

is_rstudio <- !is.na(Sys.getenv("RSTUDIO", unset = NA))
is_unix <- .Platform$OS.type == "unix"
cores <- round(parallel::detectCores() / 2)
(cores <- parallel::detectCores() - 6L)
if (!is_rstudio && is_unix) plan(multicore, workers = cores) else plan(multisession, workers = cores)


furrr::future_pmap(index_list, calc_condition_indices)
